Added Migrate and Rename buttons for SaveItem

+++ composeApp/src/commonMain/kotlin/com/davidlukash/cakebaker/ui/screens/savescreen/SaveItem.kt
@@ -2,8 +2,10 @@ package com.davidlukash.cakebaker.ui.screens.savescreen
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.FlowRow
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.defaultMinSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
@@ -14,8 +16,12 @@ import androidx.compose.runtime.getValue
 import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
+import com.davidlukash.cakebaker.VERSION
+import com.davidlukash.cakebaker.VERSIONCODE
+import com.davidlukash.cakebaker.data.Save
 import com.davidlukash.cakebaker.data.SaveFile
 import com.davidlukash.cakebaker.data.theme.Theme
+import com.davidlukash.cakebaker.data.theme.getDefaultTheme
 import com.davidlukash.cakebaker.ui.Container
 import com.davidlukash.cakebaker.ui.LargeThemedButton
 
@@ -25,6 +31,7 @@ import com.davidlukash.cakebaker.viewmodel.LocalMainViewModel
 import com.davidlukash.cakebaker.withErrorHandling
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.launch
+import org.jetbrains.compose.ui.tooling.preview.Preview
 
 @Composable
 fun SaveItem(
@@ -37,19 +44,24 @@ fun SaveItem(
 ) {
     Container(
         theme = theme,
-        modifier = Modifier.fillMaxWidth().height(360.dp)
+        modifier = Modifier.fillMaxWidth()
     ) {
         Column(
-            verticalArrangement = Arrangement.spacedBy(16.dp)
+            verticalArrangement = Arrangement.spacedBy(8.dp)
         ) {
             Text(
                 saveFile.name,
                 style = theme.buttonTextStyle,
-                modifier = Modifier.padding(8.dp)
+                modifier = Modifier.padding(horizontal = 8.dp)
             )
-            Spacer(modifier = Modifier.weight(1f))
-            Row(
+            Text(
+                "Version: ${saveFile.save.version}",
+                style = theme.labelStyle,
+                modifier = Modifier.padding(horizontal = 8.dp)
+            )
+            FlowRow(
                 horizontalArrangement = Arrangement.spacedBy(16.dp),
+                verticalArrangement = Arrangement.spacedBy(8.dp)
             ) {
                 LargeThemedButton(
                     theme = theme,
@@ -60,20 +72,35 @@ fun SaveItem(
                 ) {
                     Text("Export", style = theme.buttonTextStyle)
                 }
+                if (!saveFile.isDefault)
+                    LargeThemedButton(
+                        theme = theme,
+                        onClick = {
+                            deleteSave(saveFile)
+                        },
+                        modifier = Modifier.weight(1f),
+                    ) {
+                        Text("Delete", style = theme.buttonTextStyle)
+                    }
+                if (saveFile.save.versionCode != null && saveFile.save.versionCode <= VERSIONCODE)
+                    LargeThemedButton(
+                        theme = theme,
+                        onClick = {
+                            //exportSave(saveFile)
+                        },
+                        modifier = Modifier.weight(1f).defaultMinSize(minWidth = 600.dp)
+                    ) {
+                        Text("Migrate", style = theme.buttonTextStyle)
+                    }
                 LargeThemedButton(
                     theme = theme,
                     onClick = {
-                        deleteSave(saveFile)
+                        //deleteSave(saveFile)
                     },
                     modifier = Modifier.weight(1f),
-                    enabled = !saveFile.isDefault
                 ) {
-                    Text("Delete", style = theme.buttonTextStyle)
+                    Text("Rename", style = theme.buttonTextStyle)
                 }
-            }
-            Row(
-                horizontalArrangement = Arrangement.spacedBy(16.dp),
-            ) {
                 LargeThemedButton(
                     theme = theme,
                     onClick = {
@@ -83,17 +110,34 @@ fun SaveItem(
                 ) {
                     Text("Load", style = theme.buttonTextStyle)
                 }
-                LargeThemedButton(
-                    theme = theme,
-                    onClick = {
-                        overwriteSave(saveFile)
-                    },
-                    modifier = Modifier.weight(1f),
-                    enabled = !saveFile.isDefault
-                ) {
-                    Text("Overwrite", style = theme.buttonTextStyle)
-                }
+                if (!saveFile.isDefault)
+                    LargeThemedButton(
+                        theme = theme,
+                        onClick = {
+                            overwriteSave(saveFile)
+                        },
+                        modifier = Modifier.weight(1f).defaultMinSize(minWidth = 600.dp),
+                    ) {
+                        Text("Overwrite", style = theme.buttonTextStyle)
+                    }
             }
         }
     }
+}
+
+@Preview(widthDp = 960)
+@Composable
+fun SaveItemPreview() {
+    val theme = getDefaultTheme()
+    SaveItem(
+        theme = theme,
+        exportSave = {},
+        deleteSave = {},
+        loadSave = {},
+        overwriteSave = {},
+        saveFile = SaveFile(name = "savepreview", save = Save.default.copy(
+            version = "Unknown",
+            versionCode = 2
+        ))
+    )
 }
